api_genai_reaseguros/requirements.txt
django-environ==0.12.0
psycopg2-binary==2.9.10
django-cors-headers==4.6.0
djangorestframework-simplejwt==5.4.0
google-cloud-discoveryengine==0.13.5
google-cloud-secret-manager==2.23.1
google-cloud-bigquery==3.27.0
pydantic==2.10.5
pydantic_core==2.27.2
Django==4.2.18
google-cloud-logging==3.12.0
google-cloud-trace
requests
gunicorn==21.2.0
cloud-sql-python-connector[pg8000]==1.5.0
pg8000==1.30.3